services/viral-score-api/requirements.txt
flask==3.0.0
tensorflow==2.15.0
google-cloud-storage==2.14.0
gunicorn==21.2.0